for products having variants, our auto piloting feature is failing to move to quantity field.
alsothei make sure the variants and options inclding all products have cost , quantity and sell price as determined by the expense bill.

Also, in expense bill, show the total/sold for quantity and total show generated margin (that means, before selling the margin is negative (so show red) but as we sell, it moves to positve, completely stocking out and going outofstock.